Update skill manifest to defined intents
We considered separating "on" and "off" intents, to disambiguate "I dont want x on" and similarly
"X is too high" and possibly other cases. Also considred grouping "off" and "close" together to support
"shut off the lights" and "shut off the blinds" and "shut the locks". However, Juston wants to give an
error in those cases. So making the skill intents "on_off", "open_close", "lock_unlock" as we discussed
in person.